included a structure

+++ Dice.cpp
@@ -10,10 +10,11 @@
 
 int Dice::rollDice() { // rollDice function rolls the dice
 
-    srand((unsigned int) time(NULL)); // seeds random number generator; initializes time to NULL
-
+    srand(time(NULL)); // seeds random number generator; initializes time to NULL
     int iDicevalue = rand() % 6 + 1; // iDiceValue integer is assigned to a random number 1 through 6
 
     return iDicevalue;
 }
+
+
